refactor: extract useSplitProfileActions from Dashboard
Extract split profile CRUD operations and recipient validation into
useSplitProfileActions hook (424 lines). Removes ~370 lines from
Dashboard.js (3664 -> 3314 lines).

Hook contains: fetchSplitProfiles, saveSplitProfile, deleteSplitProfile,
setActiveSplitProfileById, validateRecipientUsername (Blink + npub.cash),
addRecipientToProfile, removeRecipientFromProfile, plus debounced
validation and fetch-on-auth useEffects.

Also removes now-unused lnurl imports from Dashboard.js.
